--- src/data/pokemon.json
@@ -1,28 +1,0 @@
-{
-    "pokemons": [
-        {
-            "id" : 4,
-            "englishname" : "Charmander",
-            "type" : "Fire"
-
-        },
-
-        {
-            "id" : 5,
-            "englishname" : "Charmelon",
-            "type" : "Fire"
-        },
-
-        {
-            "id" : 6,
-            "englishname" : "Charizard",
-            "type" : "Fire / Flying"
-        },
-        {
-            "id" : 25,
-            "englishname" : "Pikachu",
-            "type" : "Electric"
-        }
-    ]
-
-}
