+++ src/hook/useFetch.ts
@@ -1,0 +1,5 @@
+export const useFetch = async (url: string): Promise<any> => {
+   const res = await fetch(url);
+   const data = await res.json();
+   return data;
+};
