+++ src/main.tsx
@@ -1,5 +1,10 @@
-import ReactDOM from 'react-dom/client'
-import App from './App.tsx'
-import './index.css'
+import ReactDOM from "react-dom/client";
+import App from "./App.tsx";
+import "./index.css";
+import { ContextPokemon as ContextPokemonProvider } from "./context/PokemonContext.tsx";
 
-ReactDOM.createRoot(document.getElementById('root')!).render(<App />)
+ReactDOM.createRoot(document.getElementById("root")!).render(
+   <ContextPokemonProvider>
+      <App />
+   </ContextPokemonProvider>
+);
